requirements.txt
tqdm
pandas
openpyxl
SQLAlchemy
PyMySQL
pandera
cryptography
python-dotenv
